requirements.txt
streamlit==1.28.0
pandas==2.1.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.0
numpy==1.24.0
scikit-learn==1.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.4.2
python-multipart==0.0.6
openai==1.3.0
requests==2.31.0
tqdm==4.66.1 